Do not allow quitting Yaffe if it's in the background. This could happen if the emulator took foreground control back from the overlay.

+++ Yaffe/YaffeOverlay.cpp
@@ -30,7 +30,7 @@ private:
 
 	static MODAL_CLOSE(OverlayModalClose)
 	{ 
-		if (pResult == MODAL_RESULT_Ok)
+		if (pResult == MODAL_RESULT_Ok && IsApplicationFocused())
 		{
 			TerminateProcess(g_state.running_rom.hProcess, 0);
 			WaitForSingleObject(g_state.running_rom.hProcess, INFINITE);
